# ARROW-11383: [Rust] Faster bit AND and OR (2x)